% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{chat_model}
\alias{chat_model}
\alias{chat_design}
\title{Create chat model}
\usage{
chat_model(...)

chat_design(.edibble, ...)
}
\arguments{
\item{...}{Arguments parsed to the chat model.}

\item{.edibble}{An edibble experimental design object.}
}
\description{
This initiates an instance of the chat model specified
in the options. To get the current model, check \code{souschef_option_get("llm")}
and to set a different model, use \code{souschef_option_set("llm", list(vendor = "openai", model = "gpt-4o-mini"))}
replacing the vendor and vendor arguments. The model must be available in \code{elmer}.
}
