% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{describe_context}
\alias{describe_context}
\alias{describe_scenario}
\title{Describe context}
\usage{
describe_context(.edibble)

describe_scenario(.edibble)
}
\arguments{
\item{.edibble}{An edibble design.}
}
\description{
\code{describe_context()} describes the context for the given edibble experiment
in words. The \code{describe_scenario()} takes the former context and uses a LLM
to create an embellished version of the context.
}
