% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideate.R
\name{ideate}
\alias{ideate}
\alias{ideate_design}
\alias{ideate_experiment}
\alias{ideate_live_browser}
\alias{ideate_live_console}
\title{Ideate about possible improvements on the design or experiment.}
\usage{
ideate_design(.edibble)

ideate_experiment(.edibble)

ideate_live_browser(.edibble)

ideate_live_console(.edibble)
}
\arguments{
\item{.edibble}{An edibble design or edibble context.}
}
\description{
If the focus is on the design use \code{ideate_design()}. For the whole experiment, then \code{ideate_experiment()}.
}
