% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{is_edibble_context}
\alias{is_edibble_context}
\title{Check if it is an edibble context}
\usage{
is_edibble_context(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
An edibble context would be a text with the edibble design
in the attribute and optionally a chat object.
}
