% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/souschef-package.R
\docType{package}
\name{souschef-package}
\alias{souschef}
\alias{souschef-package}
\title{souschef: An AI assistant for designing edibble experiments}
\description{
Given an edibble experimental design, the large language model tries to infer the context and design simulation that may be suitable.
}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID}) [copyright holder]

}
\keyword{internal}
