% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{souschef_option_get}
\alias{souschef_option_get}
\alias{souschef_option_set}
\title{Get or set an option for the \code{souschef} package}
\usage{
souschef_option_get(x = c("llm", "all"))

souschef_option_set(x = "llm", val)
}
\arguments{
\item{x}{The name of the option to get or set.}

\item{val}{The value of the option to set.}
}
\description{
Get or set an option for the \code{souschef} package
}
