% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takeauto.R
\name{takeauto}
\alias{takeauto}
\title{Create a named experimental design with context}
\usage{
takeauto(recipe = NULL, show = TRUE, context = "")
}
\arguments{
\item{recipe}{The name of the experimental design specified as character,
or a recipe deisgn.}

\item{show}{Whether to show the code or not.}

\item{context}{A text describing the context.}
}
\description{
Similar to \code{takeout()}, this function generates a named experimental design
but makes up a context.
}
\examples{
takeauto()


}
